macos: massive reorg
