+++ macos/Sources/Ghostty/Package.swift
@@ -1,0 +1,1 @@
+struct Ghostty {}
